+++ okta/services/idaas/resource_okta_app_basic_auth.go
@@ -2,11 +2,11 @@ package idaas
 
 import (
 	"context"
+	"github.com/okta/terraform-provider-okta/okta/utils"
 	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/okta/terraform-provider-okta/okta/utils"
 	"github.com/okta/terraform-provider-okta/sdk"
 	"github.com/okta/terraform-provider-okta/sdk/query"
 )
@@ -34,6 +34,53 @@ other arguments that changed will be applied.`,
 				Type:        schema.TypeString,
 				Required:    true,
 				Description: "The URL of the sign-in page for this app.",
+			},
+			"credentials_scheme": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Default:     "EDIT_USERNAME_AND_PASSWORD",
+				Description: "Application credentials scheme. One of: `EDIT_USERNAME_AND_PASSWORD`, `ADMIN_SETS_CREDENTIALS`, `EDIT_PASSWORD_ONLY`, `EXTERNAL_PASSWORD_SYNC`, or `SHARED_USERNAME_AND_PASSWORD`",
+			},
+			"reveal_password": {
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     false,
+				Description: `Allow user to reveal password. Default is false. It can not be set to true if credentials_scheme is "ADMIN_SETS_CREDENTIALS", "SHARED_USERNAME_AND_PASSWORD" or "EXTERNAL_PASSWORD_SYNC".`,
+			},
+			"shared_username": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Shared username, required for certain schemes.",
+			},
+			"shared_password": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Shared password, required for certain schemes.",
+			},
+			"user_name_template": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Default:     "${source.login}",
+				Description: "Username template. Default: `${source.login}`",
+			},
+			"user_name_template_suffix": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Username template suffix",
+			},
+			"user_name_template_type": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Default:     "BUILT_IN",
+				Description: "Username template type. Default: `BUILT_IN`. Valid values: `NONE`, `CUSTOM`, `BUILT_IN`",
+			},
+			"user_name_template_push_status": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Push username on update. Valid values: `PUSH`, `DONT_PUSH` and `NOT_CONFIGURED`",
+				DiffSuppressFunc: func(k, old, new string, d *schema.ResourceData) bool {
+					return new == ""
+				},
 			},
 		}),
 		Timeouts: &schema.ResourceTimeout{
@@ -73,6 +120,13 @@ func resourceAppBasicAuthRead(ctx context.Context, d *schema.ResourceData, meta 
 	}
 	_ = d.Set("url", app.Settings.App.Url)
 	_ = d.Set("auth_url", app.Settings.App.AuthURL)
+	_ = d.Set("credentials_scheme", app.Credentials.Scheme)
+	_ = d.Set("reveal_password", app.Credentials.RevealPassword)
+	_ = d.Set("shared_username", app.Credentials.UserName) // We can sync shared username but not password from upstream
+	_ = d.Set("user_name_template", app.Credentials.UserNameTemplate.Template)
+	_ = d.Set("user_name_template_type", app.Credentials.UserNameTemplate.Type)
+	_ = d.Set("user_name_template_suffix", app.Credentials.UserNameTemplate.Suffix)
+	_ = d.Set("user_name_template_push_status", app.Credentials.UserNameTemplate.PushStatus)
 	appRead(d, app.Name, app.Status, app.SignOnMode, app.Label, app.Accessibility, app.Visibility, app.Settings.Notes)
 	_ = d.Set("logo_url", utils.LinksValue(app.Links, "logo", "href"))
 	return nil
@@ -125,6 +179,7 @@ func buildAppBasicAuth(d *schema.ResourceData) *sdk.BasicAuthApplication {
 		Notes: BuildAppNotes(d),
 	}
 	app.Visibility = BuildAppVisibility(d)
+	app.Credentials = BuildSchemeAppCreds(d)
 	app.Accessibility = BuildAppAccessibility(d)
 
 	return app